Move synchronous API to api module
